backend/requirements.txt
openai-agents>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
duckdb>=1.2.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.14.0
matplotlib>=3.9.0
seaborn>=0.13.0
pydantic>=2.10.0
python-dotenv>=1.0.0
